+++ packages/create-block/lib/index.js
@@ -59,6 +59,10 @@ program
 	.option( '--wp-env', 'enable integration with `@wordpress/env` package' )
 	.option( '--no-plugin', 'scaffold only block files' )
 	.option( '--variant <variant>', 'the variant of the template to use' )
+	.option(
+		'--target-dir <directory>',
+		'the directory where the files will be added'
+	)
 	.action(
 		async (
 			slug,
@@ -72,6 +76,7 @@ program
 				wpScripts,
 				wpEnv,
 				variant,
+				targetDir,
 			}
 		) => {
 			await checkSystemRequirements( engines );
@@ -102,6 +107,7 @@ program
 						title,
 						wpScripts,
 						wpEnv,
+						targetDir,
 					} ).filter( ( [ , value ] ) => value !== undefined )
 				);