+++ packages/create-block/lib/scaffold.js
@@ -2,6 +2,7 @@
  * External dependencies
  */
 const { pascalCase, snakeCase } = require( 'change-case' );
+const { join, basename } = require( 'path' );
 
 /**
  * Internal dependencies
@@ -40,6 +41,7 @@ module.exports = async (
 		npmDevDependencies,
 		customScripts,
 		folderName,
+		targetDir,
 		editorScript,
 		editorStyle,
 		style,
@@ -56,6 +58,26 @@ module.exports = async (
 ) => {
 	slug = slug.toLowerCase();
 	namespace = namespace.toLowerCase();
+	let pathToBlockFiles;
+	let rootDirectory;
+	// targetDir default is the current directory so we can rely on it.
+	if ( targetDir === './' ) {
+		rootDirectory = join( process.cwd(), targetDir, slug );
+	} else {
+		rootDirectory = plugin
+			? join( process.cwd(), targetDir, slug )
+			: join( process.cwd(), targetDir );
+	}
+
+	if ( targetDir === './' ) {
+		pathToBlockFiles = plugin
+			? join( rootDirectory, folderName )
+			: rootDirectory;
+	} else {
+		pathToBlockFiles = plugin
+			? join( rootDirectory, folderName )
+			: join( rootDirectory, folderName, slug );
+	}
 
 	const transformedValues = transformer( {
 		$schema,
@@ -81,6 +103,7 @@ module.exports = async (
 		npmDependencies,
 		npmDevDependencies,
 		customScripts,
+		targetDir,
 		folderName,
 		editorScript,
 		editorStyle,
@@ -94,6 +117,8 @@ module.exports = async (
 		customBlockJSON,
 		example,
 		textdomain: slug,
+		rootDirectory,
+		pathToBlockFiles,
 	} );
 
 	const view = {
@@ -131,7 +156,8 @@ module.exports = async (
 					await writeOutputTemplate(
 						pluginOutputTemplates[ outputFile ],
 						outputFile,
-						view
+						view,
+						rootDirectory
 					)
 			)
 		);
@@ -143,7 +169,8 @@ module.exports = async (
 				await writeOutputAsset(
 					outputAssets[ outputFile ],
 					outputFile,
-					view
+					view,
+					rootDirectory
 				)
 		)
 	);
@@ -165,8 +192,12 @@ module.exports = async (
 
 	success(
 		plugin
-			? `Done: WordPress plugin ${ title } bootstrapped in the ${ slug } directory.`
-			: `Done: Block "${ title }" bootstrapped in the ${ slug } directory.`
+			? `Done: WordPress plugin ${ title } bootstrapped in the ${ basename(
+					rootDirectory
+			  ) } directory.`
+			: `Done: Block "${ title }" bootstrapped in the ${ basename(
+					rootDirectory
+			  ) } directory.`
 	);
 
 	if ( plugin && wpScripts ) {