+++ packages/create-block/lib/templates.js
@@ -239,6 +239,7 @@ const getDefaultValues = ( pluginTemplate, variant ) => {
 		wpEnv: false,
 		npmDependencies: [],
 		folderName: './src',
+		targetDir: './',
 		editorScript: 'file:./index.js',
 		editorStyle: 'file:./index.css',
 		style: 'file:./style-index.css',